higlass check edited to look at indexing queue
check_expsets_processedfiles_for_modified_higlass_items now looks
at the indexing queue to see if items are indexing before completing,
and completes with a PASS if so. This prevents viewconfs being
created with released to lab status as items are being released,
resulting in a status mismatch after indexing is done.
Only one higlass check was modified, as it was the only one causing problems.