modify badge check for single bio rep (#389)
This check adds a badge to ExpSets with a single biological replicate. Added an exception to account for false positives.

An ExpSet might have many replicates, but only one present in the database. This happens especially with imaging datasets, where only one replicate with raw data is show as an example, but processed results are derived from many replicates. Such ExpSets are identified by the tag `many_replicates`